[Doc][Org] move edge_softmax to functional (#2442)
* move edge_softmax to functional

* change examples

* fixes

* revert

* fix

* fix

* remove

* fix

* test

* add init
