+++ examples/pytorch/gat/gat.py
@@ -10,7 +10,7 @@ Pytorch implementation: https://github.com/Diego999/pyGAT
 import torch
 import torch.nn as nn
 import dgl.function as fn
-from dgl.nn.pytorch import edge_softmax, GATConv
+from dgl.nn import GATConv
 
 
 class GAT(nn.Module):
