+++ python/dgl/nn/__init__.py
@@ -16,6 +16,11 @@ with "[NN] XXX module".
 import importlib
 import sys
 import os
+
+# [BarclayII] Not sure what's going on with pylint.
+# Possible issue: https://github.com/PyCQA/pylint/issues/2648
+from . import functional        # pylint: disable=import-self
+
 from ..backend import backend_name
 from ..utils import expand_as_pair
