+++ python/dgl/nn/functional/__init__.py
@@ -1,0 +1,3 @@
+"""Functions related to DGL NN Modules."""
+
+from ...ops import edge_softmax
