+++ python/dgl/nn/pytorch/conv/dotgatconv.py
@@ -3,7 +3,7 @@
 from torch import nn
 
 from .... import function as fn
-from ....ops import edge_softmax
+from ...functional import edge_softmax
 from ....base import DGLError
 from ....utils import expand_as_pair
