+++ python/dgl/nn/pytorch/softmax.py
@@ -1,3 +1,3 @@
 """Torch modules for graph related softmax."""
 # pylint: disable= unused-import
-from ...ops import edge_softmax
+from ..functional import edge_softmax
